requirements.txt
pyperclip~=1.8.2
keyboard~=0.13.5
future~=0.18.2
pystray~=0.19.4
Pillow~=9.2.0